Draft: Lower minimum Python interpreter version
1. Downgrade from Python 3.11 to Python 3.9

Python 3.9 is the main version for the alias `python3` in AlmaLinux 9
and `lxplus.cern.ch` nodes. Also, this decision is mainly based on the
Python version available in `cms-sw` apptainer images for `el9`, using a
different version will reduce chances to smoothly integrate this with
some batch execution within McM web application.