Use one isolated session per request
This modifies the behavior to avoid reusing the same underlying connection as it could be aborted before receiving a response.
The last raises an error users have to handle on their own, causing the user's code to stop.
If the request is not idempotent, repeating it automatically in this component could cause further unexpected issues.

An explanation of this behavior and the reason why this patch is applied is available at: https://bugs.python.org/issue41345